+++ api/graphql/mutations/savedSearch.js
@@ -1,0 +1,7 @@
+export function createSavedSearch (attributes) {
+  return SavedSearch.create(attributes)
+}
+
+export function deleteSavedSearch (savedSearchId) {
+  return SavedSearch.delete(savedSearchId)
+}